+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix converter in case of internal wallet (MINOR)
 - Fix user role order (PATCH)
 - Issue with wallet connection on the Add Wallet button has been resolved. (MINOR)
+- Fix undefined user issue in new dashboard (MINOR)
 
 # 1.3.0